P1, and a solution that works for everything _EXCEPT_ the real input.  Answer is too high...  know I'm making a silly mistake.   Just not seeing it, so stepping away for a bit.

+++ 2023/Day14.cs
@@ -51,7 +51,7 @@ namespace AoC2023
          int points = 0;
          foreach ((ivec2 pos, int val) in map) {
             if (val == 2) {
-               points += (height - pos.y);
+               points += (int)(height - pos.y);
             }
          }
